moved async storage to utils
