+++ src/app/(tabs)/home/index.tsx
@@ -12,7 +12,7 @@ import HomeSearchBar from "@components/home/HomeSearchBar";
 import { useAuth } from "@context/auth";
 import GrowToggle from "@components/home/GrowToggle";
 import { postType } from "freebites-types";
-import { setItem } from "../../../local-storage/asyncStorage";
+import { setItem } from "../../../utils/asyncStorage";
 
 const Home = () => {
   const [AllPosts, setPosts] = useState<postType[]>([]);
