+++ src/components/home/FilterList.tsx
@@ -7,7 +7,7 @@ import { FontAwesome5 } from "@expo/vector-icons";
 import { MaterialIcons } from "@expo/vector-icons";
 import { Feather } from "@expo/vector-icons";
 import { AppContext, filterTypes } from "@context/appContext";
-import { setItem } from "../../local-storage/asyncStorage";
+import { setItem } from "../../utils/asyncStorage";
 import { useBottomSheetModal } from "@gorhom/bottom-sheet";
 
 const FilterList = () => {
